split ML models and agents into separate folders, add init.py for unittes
