+++ data_discovery_ai/core/agents/linkGroupingAgent.py
@@ -4,7 +4,7 @@ from itertools import product, permutations
 import requests
 
 from data_discovery_ai import logger
-from data_discovery_ai.model.baseAgent import BaseAgent
+from data_discovery_ai.core.agents.baseAgent import BaseAgent
 from data_discovery_ai.utils.config_utils import ConfigUtil
