+++ tests/core/agents/test_keyword_classification_agent.py
@@ -1,7 +1,7 @@
 #  unit test for model/keywordClassificationAgent.py
 import unittest
 from unittest.mock import patch, MagicMock
-from data_discovery_ai.model.keywordClassificationAgent import (
+from data_discovery_ai.core.agents.keywordClassificationAgent import (
     KeywordClassificationAgent,
 )
 from data_discovery_ai.utils.config_utils import ConfigUtil
